+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # @watergis/maplibre-gl-terradraw
 
+## 1.15.2
+
+### Patch Changes
+
+- 9223f70: Applied delete confirmation functionality to deleting selected features too. It was only available when deleting all features on the map
+
+## 1.15.1
+
+### Patch Changes
+
+- 75a40a8: chore: upgrade terra-draw to v1.32.2
+- 63e8eb4: chore: support maplibre-gl v6
+
 ## 1.15.0
 
 ### Minor Changes